+++ content_editor/static/content_editor/save_shortcut.js
@@ -1,9 +1,11 @@
-/* global django */
-django.jQuery(($) => {
-  $(document).keydown(function handleKeys(event) {
-    if (event.which === 83 && (event.metaKey || event.ctrlKey)) {
-      $("form input[name=_continue]").click()
-      return false
-    }
-  })
+window.addEventListener("keydown", (e) => {
+  let el
+  if (
+    e.key === "s" &&
+    (e.metaKey || e.ctrlKey) &&
+    (el = document.querySelector("form input[name=_continue]"))
+  ) {
+    e.preventDefault()
+    el.click()
+  }
 })